Fix name patterns (regular expression) for packages and resources
Characters like for example the single character match '.' do not
have to be escaped in bracket expression because within the scope
of bracket expressions they match the literal characters. There are
also unknown "escape sequences" in the bracket expression like for
example \_. Therefore removing any escapes of characters in bracket
expressions for names of packages and resources in both data
package schema and tabular data package schema.