Add fiscal data package schema
The fiscal data package schema relies heavily on $ref which for now
only points to raw github uris but should optimally be a nicer
domain, e.g. schemas.dataprotocols.org

The schema is compatible with version 0.3.0-alpha2 of fiscal data
package and tries as hard as possible to tackle the mapping
property which allows random dimension names and measures names
which require some patternProperties hacking. Field primary key for
dimensions is also not properly addressed in the schema because it
has to be present in at least one field, but not in all. That's
very difficult to address in a schema.

There are also inconsistencies in the spec, e.g. around dimension
types which are listed in an example and in a list of possible
values and the two lists are inconsistent. We go with the spec
list rather than the listed options in the example in the spec.

There are some pull requests open on the fiscal data package spec
which will require changes to this schema, but as is, it's usable
for the most part.